requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.48
alembic==1.13.3
psycopg[binary]==3.3.3

# Settings & environment
pydantic-settings==2.5.2
email-validator==2.2.0
python-dotenv==1.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client (for TMDB API)
httpx==0.27.2

# Google Gemini AI
google-genai==1.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # also used by TestClient
